Implemented set of analog filters (Butterworth, Chebyshev, ChebyshevII)

+++ src/main/scala/com/sparja/syto/filter/butterworth/DigitalButterworthHighPassFilter.scala
@@ -3,7 +3,7 @@ package com.sparja.syto.filter.butterworth
 import com.sparja.syto.filter.core.{HighPassTransferFunction, TransferFunction}
 
 object DigitalButterworthHighPassFilter {
-  def apply(order: Int, cutoffFrequency: Float, sampleFrequency: Float): TransferFunction = {
+  def apply(order: Int, cutoffFrequency: Double, sampleFrequency: Double): TransferFunction = {
     val approximation = new ButterworthApproximation(order)
     new HighPassTransferFunction(approximation, order, cutoffFrequency, sampleFrequency)
   }
