+++ src/main/scala/com/sparja/syto/filter/chebyshev/DigitalChebyshevFirstTypeLowPassFilter.scala
@@ -3,7 +3,7 @@ package com.sparja.syto.filter.chebyshev
 import com.sparja.syto.filter.core.{LowPassTransferFunction, TransferFunction}
 
 object DigitalChebyshevFirstTypeLowPassFilter {
-  def apply(order: Int, rp: Float, cutoffFrequency: Float, sampleFrequency: Float): TransferFunction = {
+  def apply(order: Int, rp: Double, cutoffFrequency: Double, sampleFrequency: Double): TransferFunction = {
     val approximation = new ChebyshevFirstTypeApproximation(order, rp)
     new LowPassTransferFunction(approximation, order, cutoffFrequency, sampleFrequency)
   }
