+++ src/main/scala/com/sparja/syto/filter/core/TransferFunction.scala
@@ -2,6 +2,6 @@ package com.sparja.syto.filter.core
 
 trait TransferFunction {
 
-  def calculateCoefficients(): (List[Float], List[Float])
+  def calculateCoefficients(): (List[Double], List[Double])
 
 }
